Some more test coverage
